adding test conversions in modal dir
