+++ packages/augur-ui/src/modules/modal/actions/close-modal.test.js
@@ -1,0 +1,13 @@
+import { CLOSE_MODAL, closeModal } from "modules/modal/actions/close-modal";
+
+describe("modules/modal/actions/close-modal", () => {
+  test("should return the expected value", () => {
+    const actual = closeModal();
+
+    const expected = {
+      type: CLOSE_MODAL
+    };
+
+    expect(actual).toEqual(expected);
+  });
+});
